requirements.txt
flask
gunicorn
requests
beautifulsoup4
lxml
psycopg2-binary
